Updated for latest grblHAL version - 20250626
Changed to use the grblHAL RGB HAL/API.
Fix for potential violent stop during execution of feed hold since updating LEDs is done with interrupts disabled which blocks step generation. This is more likely when long strips are configured.
Fix for potential alarn 14 beeing raised if LEDs are updated when a Modbus transaction is ongoing and a Modbus controlled VFD is used. This is due to disabled interrupts leading to corrupted or missed messages.
Removed files and code no longer needed++